flask_deepface_emotion_detection/requirements.txt
flask
flask-cors
tf-keras
deepface
opencv-python-headless
Pillow
numpy
gunicorn